backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
sqlalchemy>=2.0.0
pydantic>=2.6.0
python-dotenv>=1.0.0
httpx>=0.27.0
apscheduler>=3.10.0
pdfplumber>=0.11.0
python-docx>=1.1.0
python-multipart>=0.0.9
